backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pytesseract
pdf2image
Pillow
groq
requests
aiofiles
starlette
python-dotenv
numpy
